updating to java 21 and springboot 3.5.6
