+++ src/test/java/uk/gov/onsdigital/banner/ControllerAdviceIntegrationTest.java
@@ -6,9 +6,9 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.RequestBuilder;
@@ -18,15 +18,15 @@ import uk.gov.onsdigital.banner.service.BannerService;
 import uk.gov.onsdigital.banner.service.TemplateService;
 
 @WebMvcTest(value = TemplateController.class)
-public class ControllerAdviceIT {
+public class ControllerAdviceIntegrationTest {
   
   @Autowired
   private MockMvc mockMvc;
 
-  @MockBean
+  @MockitoBean
   private TemplateService templateService;
 
-  @MockBean
+  @MockitoBean
   private BannerService bannerService;
