+++ src/test/java/uk/gov/onsdigital/banner/controller/TemplateControllerIntegrationTest.java
@@ -1,7 +1,5 @@
 package uk.gov.onsdigital.banner.controller;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import java.net.URI;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -10,17 +8,22 @@ import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import uk.gov.onsdigital.banner.DatastoreEmulator;
-import uk.gov.onsdigital.banner.model.BannerModel;
-import uk.gov.onsdigital.banner.service.BannerService;
+import uk.gov.onsdigital.banner.model.TemplateModel;
+import uk.gov.onsdigital.banner.service.TemplateService;
+
+import java.net.URI;
+import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-public class BannerControllerIT {
+public class TemplateControllerIntegrationTest {
 
   @BeforeAll
   public static void setUpClass() {
@@ -38,30 +41,41 @@ public class BannerControllerIT {
   @Autowired
   private TestRestTemplate restTemplate;
 
-  @MockBean
-  private BannerService bannerService;
+  @MockitoBean
+  private TemplateService templateService;
+
+  @Test
+  public void willReturnBannersObject() {
+    Mockito.when(templateService.getAllTemplates())
+        .thenReturn(List.of(TemplateModel.builder().build()));
+    List<TemplateModel> banners = this.restTemplate.getForObject("http://localhost:" + port + "/template",
+        List.class);
+
+    assertEquals(1, banners.size());
+  }
 
   @Test
   public void willReturnASingleBannerObject() {
-    BannerModel expectedBanner = BannerModel.builder().build();
-    Mockito.when(bannerService.getBanner("active"))
+    TemplateModel expectedBanner = TemplateModel.builder().build();
+    Mockito.when(templateService.getTemplate("1"))
         .thenReturn(expectedBanner);
-    BannerModel actualBanner = this.restTemplate.getForObject("http://localhost:" + port + "/banner",
-        BannerModel.class);
+    TemplateModel actualBanner = this.restTemplate.getForObject("http://localhost:" + port + "/template/1",
+        TemplateModel.class);
 
     assertEquals(expectedBanner, actualBanner);
   }
 
   @Test
   public void willCreateBanner() {
-    BannerModel postedBanner = BannerModel.builder()
+    TemplateModel postedBanner = TemplateModel.builder()
+        .title("BannerTitle")
         .content("Banner Content")
         .build();
 
-    Mockito.when(bannerService.createBanner(ArgumentMatchers.any()))
+    Mockito.when(templateService.createTemplate(ArgumentMatchers.any()))
         .thenReturn(postedBanner);
-    ResponseEntity<BannerModel> resp = this.restTemplate.postForEntity("http://localhost:" + port + "/banner",
-        postedBanner, BannerModel.class);
+    ResponseEntity<TemplateModel> resp = this.restTemplate.postForEntity("http://localhost:" + port + "/template",
+        postedBanner, TemplateModel.class);
 
     assertEquals(HttpStatus.CREATED, resp.getStatusCode());
     assertEquals(postedBanner, resp.getBody());
@@ -69,8 +83,8 @@ public class BannerControllerIT {
 
   @Test
   public void willRemoveBanner() {
-    this.restTemplate.delete(URI.create("http://localhost:" + port + "/banner"));
+    this.restTemplate.delete(URI.create("http://localhost:" + port + "/template/1"));
 
-    Mockito.verify(bannerService).removeBanner("active");
+    Mockito.verify(templateService).removeTemplate("1");
   }
 }
